Improve the app, especially playing against Sensei.

+++ app/src/main/java/main/UI.java
@@ -42,5 +42,5 @@ public interface UI {
   FileAccessor fileAccessor();
   boolean active();
   boolean useBook();
-  int getError();
+  double getError();
 }
